Removed loose vendor, moved under migplan package
