+++ Gopkg.lock
@@ -149,14 +149,6 @@
   version = "v2.9.3"
 
 [[projects]]
-  digest = "1:f9a456e75f573aa4a51bfae6ae4ad238eaaaf327c442c87062ab9d002aa071ff"
-  name = "github.com/evanphx/json-patch"
-  packages = ["."]
-  pruneopts = "T"
-  revision = "bf22ed9311622d93e213ba31e4ae7a5771e5d379"
-  version = "v4.6.0"
-
-[[projects]]
   digest = "1:2cd7915ab26ede7d95b8749e6b1f933f1c6d5398030684e6505940a10f31cfda"
   name = "github.com/ghodss/yaml"
   packages = ["."]
@@ -1056,7 +1048,6 @@
   packages = [
     "discovery",
     "dynamic",
-    "dynamic/fake",
     "kubernetes",
     "kubernetes/scheme",
     "kubernetes/typed/admissionregistration/v1alpha1",
@@ -1100,7 +1091,6 @@
     "rest",
     "rest/watch",
     "restmapper",
-    "testing",
     "third_party/forked/golang/template",
     "tools/auth",
     "tools/cache",
@@ -1334,7 +1324,6 @@
     "k8s.io/apiserver/pkg/storage/names",
     "k8s.io/client-go/discovery",
     "k8s.io/client-go/dynamic",
-    "k8s.io/client-go/dynamic/fake",
     "k8s.io/client-go/kubernetes",
     "k8s.io/client-go/kubernetes/scheme",
     "k8s.io/client-go/kubernetes/typed/core/v1",
